Address CodeRabbit PR feedback improvements
✅ Fixed Issues:
- Remove unused PathBuf import from build.rs
- Improve error handling with expect() instead of unwrap()
- Fix markdown URL formatting in CLAUDE.md
- Consolidate framework linking (remove duplication from .cargo/config.toml)
- Add comprehensive documentation to WhisperBridge.h with memory management notes
- Improve error handling in build script with platform compatibility checks
- Use conditional compilation (#if RUST_FFI_ENABLED) instead of TODO comments
- Add detailed documentation to Rust transcription function

🔧 Technical Improvements:
- Change audio_len parameter from int to size_t for better type safety
- Update Rust FFI to use libc::size_t consistently
- Add ignored test case structure for future real transcription testing
- Enhance build script with explicit error checking and exit codes
- Add memory ownership documentation for WhisperResult structure

🧪 Testing:
- All builds pass: Rust (release) ✅ Swift (debug) ✅
- FFI safety tests continue to pass
- Library linking correctly resolved

+++ Sources/WhisperNode/Bridge/WhisperBridge.h
@@ -2,6 +2,7 @@
 #define WhisperBridge_h
 
 #include <stdbool.h>
+#include <stddef.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -11,17 +12,18 @@ extern "C" {
 typedef struct WhisperHandle WhisperHandle;
 
 // FFI-safe result structure matching Rust definition
+// Note: caller is responsible for freeing 'text' and 'error' using whisper_free_string()
 typedef struct {
     bool success;
-    char* text;
-    char* error;
+    char* text;   // Transcribed text (NULL if success=false)
+    char* error;  // Error message (NULL if success=true)
 } WhisperResult;
 
 // Initialize whisper context with model path
 WhisperHandle* whisper_init(const char* model_path);
 
 // Transcribe audio data (f32 array, length)
-WhisperResult whisper_transcribe(WhisperHandle* handle, const float* audio_data, int audio_len);
+WhisperResult whisper_transcribe(WhisperHandle* handle, const float* audio_data, size_t audio_len);
 
 // Free whisper context
 void whisper_free(WhisperHandle* handle);
